% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desirabilityPlot.R
\name{desirabilityPlot}
\alias{desirabilityPlot}
\title{Plot desirability profile}
\usage{
desirabilityPlot(functions, plotData, bestValues, desirab, types)
}
\arguments{
\item{functions}{An array of functions}

\item{plotData}{A data frame generated from an optimization function}

\item{bestValues}{The optimal mixture composition to be presented}

\item{desirab}{An array of desirability functions}

\item{types}{An array of strings containing the attributes to be shown for each desirability function. Currently only accepts \code{"max"} or \code{"min"}}
}
\value{
A ggplot composite object from patchwork
}
\description{
This function creates a graphical representation of the desirability profiles
within the data range. It requires a data frame object generated from
an optimization function. You can use a simpler data frame to plot the lines and
present a more accurate data from another optimization call.
}
