% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureFineOptim.R
\name{mixtureFineOptim}
\alias{mixtureFineOptim}
\title{Performs a restrict interval optimization}
\usage{
mixtureFineOptim(
  functions,
  desirabilityModel,
  startPoint,
  step = 0.001,
  alpha = 0.02
)
}
\arguments{
\item{functions}{An array of functions}

\item{desirabilityModel}{A desirability overallD model}

\item{startPoint}{An array with the references (mid-points) for the optimization}

\item{step}{The ammount of each increment in the optimization}

\item{alpha}{Defines the range of the seach, as startPoint +- alpha for each x value}
}
\value{
A list containg the data regarding the maximum desirability found
}
\description{
This function performs a optimization testing within an interval defined
by the user using starting point and an alpha value. Since it is designed
for more accurate searching, it does not allow the generation of the
data frame for plotting.
}
