% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureOptim.R
\name{mixtureOptim}
\alias{mixtureOptim}
\title{Performs a full interval optimization}
\usage{
mixtureOptim(functions, desirabilityModel, xCount, step = 0.01, plot = T)
}
\arguments{
\item{functions}{An array of functions}

\item{desirabilityModel}{A desirability overallD model}

\item{xCount}{The amount of x variables used in the functions}

\item{step}{The ammount of each increment in the optimization}

\item{plot}{Define is the data frame that can be used for the desirabilityPlot function will be create. Strongly affects performance}
}
\value{
A list containg the data regarding the maximum desirability found
}
\description{
This function performs a full interval optimization (0-1 on each x variable).
It allows the creation of the data frame used for plotting.
}
