% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureRangeOptim.R
\name{mixtureRangeOptim}
\alias{mixtureRangeOptim}
\title{Performs a specific range optimization}
\usage{
mixtureRangeOptim(
  funcoes,
  modeloDesej,
  midPoints,
  alpha,
  passo = 0.01,
  plot = T
)
}
\arguments{
\item{alpha}{Defines the range of the seach, as startPoint +- alpha for each x value}

\item{functions}{An array of functions}

\item{desirabilityModel}{A desirability overallD model}

\item{startPoint}{An array with the references (mid-points) for the optimization}

\item{step}{The ammount of each increment in the optimization}
}
\value{
A list containg the data regarding the maximum desirability found
}
\description{
This function performs a optimization testing within an interval defined
by the user using alpha values for each middle point. It allows the
generation of the data frame for plotting.
}
